Feat: WIP Home

+++ MapleServer2/Tools/FieldManagerFactory.cs
@@ -13,7 +13,7 @@ namespace MapleServer2.Tools
             Managers = new Dictionary<int, List<FieldManager>>();
         }
 
-        public FieldManager GetManager(int key, int instanceId)
+        public FieldManager GetManager(int key, long instanceId)
         {
             lock (Managers)
             {
@@ -35,7 +35,7 @@ namespace MapleServer2.Tools
             }
         }
 
-        public bool Release(int key, int instanceId)
+        public bool Release(int key, long instanceId)
         {
             lock (Managers)
             {
